fix: move TransferElementID functions to right place
relevant if TransferElement.h not included
